usb: ehci-msm-hsic: Remove usage of MSM_GPIO_TO_INT
MSM_GPIO_TO_INT is a platform-specific macro that should only
be used in board files, not drivers. In this case, we actually
don't need the GPIO number in the driver itself and only need
to manipulate it using the IRQ number it maps to, so perform the
mapping in the device file.

Change-Id: I99eb964f3c7c4e90a4d75d16b2b29c687c3437ca
Signed-off-by: Jack Pham <jackp@codeaurora.org>

Conflicts:
	drivers/usb/host/ehci-msm-hsic.c

NOTE:

Adding this commit because I am trying to match the set of
commits that elkay pulled.  However, this commit is really
not complete because there are other GPIO numbers that HTC
has added that weren't corrected via this patch.