cm3629: remove power button pocket mode check
I'm trying to track down a problem where the proximity sensor
seems to get left on indefinitely and this is one area that looks
suspicious.

While conceptually this sounds like a good check, it doesn't seem
to really work.  I put the phone in my pocket, in the middle
of the night, in a room with no lights on and pushed the power
button and the screen came on.

I suspect that this is unreliable because the pocket mode check
toggles the proximity sensor on and off very quickly and it may
take a little time for it to read reliably?