Take mdeejay's overclocking
Reverted dsb's overclocking and applied this to see if it solves
a panic related to L2 cache corruption.

This only has overclocking to 1.7GHz but it feels at least as
responsive as the 1.8GHz from DSB.  It also seems like it should
be more stable although that is hard to gauge without a lot of testing
(I have only had 1 crash in almost a month related to L2 problems).