thermal: delay triggered work until after thermal device is updated
Previously we would (potentially in parallel) call the work function
when a temperature trigger was hit.

This can cause the work function to completely execute before the
IRQ handler has finished and, most importantly, before it can
finish manipulating the thermal device's config.

If this happens we potentially clear a trigger that was set in the
work function.

Now, record the sensors that were triggered and then schedule the
work after we are done all of our internal device manipulation.