simple_plug: handle external changes to running cores
This commit makes simple_plug be more friendly to apps that insist
on starting or stopping cores on their own.

As soon as we discover that the number of running cores were externally
modified, we give a 2 minute grace people to let the application do
what it needs to do.  After the 2 minute grace period, we force the
state of the cores back to what we think it should be.

During the grace period, we make no further modification to the
state of the cores.