Set the default GPU speed to be 200000000
For some reason this seems to let it scale better.  When left at
32000000, it doesn't seem to want to go lower than that for me.

While this doesn't look believable in the code, setting the lower
default speed should in theory have no negative impact as the scaling
will change it as needed.

This also matches the setting of the M7 kernel.