monarudo-display: improve the gpio / display handling
* Pull repeated code into a common helper function to make things
  clearer (backlight_gpio_enable)
* Add locking around the display operations.  This shouldn't be
  required but the performance impact is neglible and it can protect
  us if ther are race conditions turning on/off/backlight.
* Add a sleep of 1ms after turning the backlight gpio on/off.
  This matches the gpio on code of the M7

This combination of changes fixes a problem where the backlight
doesn't always get set after moving to the correct display_on
handling (previously it would occasionally end up set to a very
low backlight).

These changes and the related changes based on the M7 kernel may
or may not help improve the flickering issue.  It is too early
to tell.