monarudo-display: enable hardware vsync support
This commit is based on the msm-8960 kernels which provide our
VSYNC_GPIO to mipi_dsi and configure the panel to use it.

Small scale testing seems to indicate that this reduces the
severity of flickering on the screen but doesn't solve it
at all.

This might increase flickering.  I ran for a day with this change
and another commit and my flickering was worse.  Even if it does
increase flickering, this is important because we don't want to
disable the clock control!