init.d: cpufreq: Set governor before min/max freqs
There is a race condition when starting up CM 11 after a clean
wipe and when the user has specified cpufreq control files.

My script can set the min, the the performance governor (the default)
to set the min to max frequency before I place a lock on the settings.

This causes the phone to be stuck at max frequency until the
next time you reboot (or use adb to manipulate /sys files).